refactor: 댓글 관련 컴포넌트 features 레이어로 이동
