+++ apps/web/src/features/comment/ui/index.ts
@@ -1,2 +1,2 @@
 export * from './CommentForm';
-export * from './CommentActionMenu';
+export * from './CommentItem';
