+++ apps/web/src/features/report/ui/index.ts
@@ -1,2 +1,1 @@
-export * from './ReportConfirmModal';
-export * from './ReportReasonModal';
+export * from './ReportFlow';
