+++ apps/web/src/widgets/post/ui/index.ts
@@ -1,5 +1,3 @@
 export * from './PostContent';
-export * from './CommentItem';
 export * from './CommentList';
-export * from './ReplyList';
 export * from './PostDetailSection';
