--- apps/web/src/widgets/report/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './ReportFlow';
